requirements.txt
# Moltbook Watchtower — minimal deps; prefer stdlib where possible
requests>=2.28.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pytest>=7.0.0,<9.0.0
